fix(data-handling): fix turned_off/1 usage in cumulated_time
there was an error in the automate flow so that conditions
were inverted. it was then fixed after March 28th, but data
were still there. We can consider 1st of April to be the first
day to have correct on_off values.
this commit fixes usage of `turned_off/1` to comply
with the new automate flow, to show correct usage data.