fix(data-handling): add fake midnight entries when loading data
"midnight entries" are necessary to calculate cumulated usage time
correctly. they replicate the same on_off value of the last entry
of the previous day, but at midnight of the current day.
it lets us calculate usage time if the screen was turned on
before midnight and was turned off after, and vice versa.